// web/src/types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
};

export type AuthenticationResult = {
  __typename?: 'AuthenticationResult';
  accessToken: Scalars['String'];
  me?: Maybe<User>;
  refreshToken: Scalars['String'];
};

/** A comic book or trade paperback */
export type Book = {
  __typename?: 'Book';
  ageRating?: Maybe<Scalars['String']>;
  alternateTitle?: Maybe<Scalars['String']>;
  createdAt: Scalars['DateTime'];
  createdBy: User;
  credits?: Maybe<Array<Credit>>;
  displayName: Scalars['String'];
  format?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  issue?: Maybe<Scalars['String']>;
  pageCount?: Maybe<Scalars['String']>;
  price?: Maybe<Scalars['String']>;
  publicationDate?: Maybe<Scalars['String']>;
  publisher?: Maybe<Publisher>;
  publisherName?: Maybe<Scalars['String']>;
  series?: Maybe<Series>;
  seriesName?: Maybe<Scalars['String']>;
  summary?: Maybe<Scalars['String']>;
  updatedAt: Scalars['DateTime'];
  updatedBy: User;
};
// ...
export type Creator = {
  __typename?: 'Creator';
  createdAt: Scalars['DateTime'];
  credits: Array<Credit>;
  firstName?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  lastName?: Maybe<Scalars['String']>;
  updatedAt: Scalars['DateTime'];
};

/** An association of book and creator */
export type Credit = {
  __typename?: 'Credit';
  book: Book;
  createdAt: Scalars['DateTime'];
  creator: Creator;
  featured: Scalars['Boolean'];
  id: Scalars['ID'];
  /** The position this creator should be listed in a featured list */
  position?: Maybe<Scalars['Int']>;
  role: Scalars['String'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type Publisher = {
  __typename?: 'Publisher';
  bookCount: Scalars['Int'];
  createdAt: Scalars['DateTime'];
  id: Scalars['ID'];
  name: Scalars['String'];
  series: Array<Series>;
  seriesCount: Scalars['Int'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type Series = {
  __typename?: 'Series';
  bookCount: Scalars['Int'];
  /** A list of books released in this series */
  books?: Maybe<Array<Book>>;
  createdAt: Scalars['DateTime'];
  id: Scalars['ID'];
  name: Scalars['String'];
  publisher?: Maybe<Publisher>;
  publisherName?: Maybe<Scalars['String']>;
  updatedAt: Scalars['DateTime'];
};
// ...
export enum SortDirectionEnum {
  /** Ascending */
  Asc = 'ASC',
  /** Descending */
  Desc = 'DESC'
}
// ...
export type User = {
  __typename?: 'User';
  avatar: Scalars['String'];
  createdAt: Scalars['DateTime'];
  email: Scalars['String'];
  id: Scalars['ID'];
  roles: Array<Scalars['String']>;
  updatedAt: Scalars['DateTime'];
  username: Scalars['String'];
};